Removing implementation details from gemm.h
